updated messenger & tests to use new rules
updated the model for Rules to always return false to  .is_satisfied? unless the rule is activated

changed the joiner logic so that a message's content is set to the key_advice field of the rule.
if you only made it reference-through to the rule, then you erase the log of what was actually sent if you update the rule